Changed SyncCallback so it uses a long for timeout.
That will make it easier to overload the constructor to take an int
with the number of expected calls.

Test: atest AdServicesTestUtilityTests
Bug: 337014024

+++ adservices/tests/unittest/service-core/src/com/android/adservices/service/common/AdServicesCommonServiceImplTest.java
@@ -1007,14 +1007,14 @@ public class AdServicesCommonServiceImplTest extends AdServicesExtendedMockitoTe
     private static final class SyncIAdServicesCommonCallback
             extends IntFailureSyncCallback<IsAdServicesEnabledResult>
             implements IAdServicesCommonCallback {
-        private SyncIAdServicesCommonCallback(int timeoutMs) {
+        private SyncIAdServicesCommonCallback(long timeoutMs) {
             super(timeoutMs);
         }
     }
 
     private static final class SyncIUpdateAdIdCallback extends IntFailureSyncCallback<String>
             implements IUpdateAdIdCallback {
-        private SyncIUpdateAdIdCallback(int timeoutMs) {
+        private SyncIUpdateAdIdCallback(long timeoutMs) {
             super(timeoutMs);
         }
     }
@@ -1022,7 +1022,7 @@ public class AdServicesCommonServiceImplTest extends AdServicesExtendedMockitoTe
     private static final class SyncIEnableAdServicesCallback
             extends IntFailureSyncCallback<EnableAdServicesResponse>
             implements IEnableAdServicesCallback {
-        private SyncIEnableAdServicesCallback(int timeoutMs) {
+        private SyncIEnableAdServicesCallback(long timeoutMs) {
             super(timeoutMs);
         }
     }
@@ -1030,7 +1030,7 @@ public class AdServicesCommonServiceImplTest extends AdServicesExtendedMockitoTe
     private static final class SyncIAdServicesCommonStatesCallback
             extends IntFailureSyncCallback<AdServicesCommonStatesResponse>
             implements IAdServicesCommonStatesCallback {
-        private SyncIAdServicesCommonStatesCallback(int timeoutMs) {
+        private SyncIAdServicesCommonStatesCallback(long timeoutMs) {
             super(timeoutMs);
         }
     }
